requirements.txt
aiogram==3.15.0
yfinance==0.2.50
pandas==2.2.3
sqlalchemy==2.0.36
loguru==0.7.3
python-dotenv==1.0.1
aiohttp==3.11.10